% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_annotated_block.R
\name{normalize_annotated_block}
\alias{normalize_annotated_block}
\title{normalize an annotated block}
\usage{
normalize_annotated_block(text)
}
\arguments{
\item{text}{the text to replace}
}
\description{
normalize an annotated block
}
